Restructure file layouts, remove comments, class to className error
