--- src/admin_pages/Admin_Outreach.js
@@ -1,29 +1,0 @@
-/// summary
-///   Not used currently - will be updated iteration 3
-/// summary
-
-import React from 'react';
-// import { Link } from 'react-router-dom'; not used
-
-// components
-import Nav from '../shared_comps/Admin_nav';
-import Footer from '../components_en/Footer_en'; 
-
-//css
-import '../css/main.css'
-
-
-
-function Admin_Outreach_Page() {
-
-
-  return(
-    <div>
-       <Nav/>
-       <Footer/>
-    </div>
-
-)
-}
-
-export default Admin_Outreach_Page;
